BInt

+++ src/p16e1.rs
@@ -260,8 +260,8 @@ impl crate::RawPosit for P16E1 {
     type UInt = u16;
     type Int = i16;
 
-    const BITSIZE: usize = 16;
+    const BITSIZE: u32 = 16;
 
-    const EXPONENT_BITS: usize = 1;
+    const EXPONENT_BITS: u32 = 1;
     const EXPONENT_MASK: Self::UInt = 0x1;
 }
