+++ src/p8e0.rs
@@ -201,8 +201,8 @@ impl crate::RawPosit for P8E0 {
     type UInt = u8;
     type Int = i8;
 
-    const BITSIZE: usize = 8;
+    const BITSIZE: u32 = 8;
 
-    const EXPONENT_BITS: usize = 0;
+    const EXPONENT_BITS: u32 = 0;
     const EXPONENT_MASK: Self::UInt = 0x0;
 }
